refactor: remove test.rs file and move tests to a module files
